+++ src/PluginModules/UpdateChecker.lua
@@ -2,15 +2,24 @@ local RunService = game:GetService("RunService")
 
 ---
 
-local PluginModules = script.Parent
+local root = script.Parent.Parent
+
+local includes = root:FindFirstChild("includes")
+local Promise = require(includes:FindFirstChild("Promise"))
+
+local PluginModules = root:FindFirstChild("PluginModules")
+local PluginEnums = require(PluginModules:FindFirstChild("PluginEnums"))
+local PluginSettings = require(PluginModules:FindFirstChild("PluginSettings"))
 local ReleaseVersion = require(PluginModules:FindFirstChild("ReleaseVersion"))
 local Translator = require(PluginModules:FindFirstChild("Translator"))
 
 ---
 
 local ASSET_ID = 6474565567
+local AUTO_CHECK_INTERVAL = 5 * 60 -- in seconds
 
 local sessionNotificationShown = false
+local autoCheckPromise
 
 ---
 
@@ -47,5 +56,32 @@ UpdateChecker.Check = function()
         sessionNotificationShown = true
     end
 end
+
+UpdateChecker.SetupAutoCheck = function()
+    if ((not RunService:IsEdit()) or autoCheckPromise) then return end
+
+    autoCheckPromise = Promise.delay(AUTO_CHECK_INTERVAL):andThen(function()
+        autoCheckPromise = nil
+
+        UpdateChecker.Check()
+        UpdateChecker.SetupAutoCheck()
+    end)
+end
+
+---
+
+PluginSettings.SettingChanged:subscribe(function(setting)
+    local key, newValue = setting.Key, setting.Value
+    if (key ~= PluginEnums.PluginSettingKey.AutoCheckForUpdate) then return end
+
+    if (newValue) then
+        UpdateChecker.SetupAutoCheck()
+    else
+        if (autoCheckPromise) then
+            autoCheckPromise:cancel()
+            autoCheckPromise = nil
+        end
+    end
+end)
 
 return UpdateChecker